Fixed unit tests.
I needed to change some test inputs because the variables are stored as
int even if the context is different. The decompilation engine needs to
support context based types.

ATM 19 test cases are not implemented but the original decompiler fails
these cases as well so I consider the rust decompiler on par with the
native one.