Added layer for the PPE target.
The PPE target is not quite PPL, it's more primitive - it makes sense to transform the PPL ast to the PPE one.
PPE is an easy stack based machine language. The execution engine should run on that it should be faster and some PPE cases are handled better on this level.

Another point is that certain things can be shared bewteen compiler & decompiler like AST optimizations, disassembler output.