Fixed decompiler tests.
Decompiler needs more work but need to investigate the decompilation issue better. The decompiler tries to follow the control flow.
Maybe there were some kind of tools trying to prevent decompiling with obfuscation. The PPE format allows to jump to arbitrary positions that opens room for obfuscators tempering with the format.