Fixed user birthday variable type.
According to the docs it should be a DATE:
U_BIRTHDATE VARIABLE (3.40)
Type & Value
DATE  The current users birthdate

But the 3.40 compiler generates a String. I don't suppose that a 3.40 final will be released so I changed
it to string as well for 3.40. In the example they give for the U_BIRTHDATE variable they're using a string:

GETUSER
U_BIRTHDATE="02-13-77"
PUTUSER()

So it may be intended.