Fixed U_BIRTHDATE type.
I decided that this is a bug in the 3.4 beta implementation and that the docs are correct.
Doesn't make sense to define a date without the date type - when comparing/setting dates it only makes problems if that is no date.

The user_var recognition works with 'wrong' input files but the compiler corrects the output.

So the U_BIRTHDATE is incompatible with the 3.40 beta release - however in the case of a future BBS that will use this compiler
or engine this behavior makes more sense.