Implemented user.inf support.
user.inf is a generic data structure containing meta information for the users.
Not sure why they split up the user records at that time.

Maybe they thought to optimize something but due to the nature of the users.inf it's likely that both users and users.inf files need to be written every time the user logs off.